backend/requirements.txt
fastapi
httpx
requests
uvicorn
aiofiles
redis-om
pandas
sentence-transformers
